BSS2-2389 modify GitHub Actions to only build docker image when CI / CD stage passes (#22)
<!-- markdownlint-disable-next-line first-line-heading -->
## What?

Only build docker image when CI / CD stage passes on feature branches.

## Why?

We only want to build the docker image on feature branches (not PRs),
and to do so when the CI / CD checks have passed, in order to make
efficient use of resources and enforce code quality checks.

## How?

Create two new top level GitHub Actions workflows to manage which
downstream workflows to run on:

1. Push to any branch
2. A PR is opened, reopened, or has a commit pushed to it

## Type of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [ ] Refactoring (non-breaking change)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [x] I am familiar with the [contributing
guidelines](https://github.com/NHSDigital/bs-select-playwright/blob/main/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [ ] I have added tests to cover my changes (where appropriate)
- [ ] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.