BSS2-2403 run tests marked with branch specific markers (#25)
## What?

Receive the branch name and insert this in to a marker template before
passing it to pytest

## Why?

Sometimes we want to run a specific set of tests when we are developing
new features.

## How?

The branch name is extracted from the front of the markers argument and
inserted into the marker template: `branch(serial='<branch name>')`.
This marker will run any tests that are marked as such:
`@pytest.mark.branch(serial='<branch name>')`

## Notes

- Full stops are now used instead of commas as the separator in the
markers argument, eg. `bss2-2403.or.smoke`, because:
1. Spaces present problems when passed as args so we have to use
something else.
2. They are safe characters in docker container file paths and S3 bucket
paths; the markers arg is used as a directory name.
3. They are not allowed characters in python method names and so can be
safely removed without checks, as they won't be present in any of the
bodies of the markers themselves.

- [Corresponding merge
request](https://gitlab.mgmt.texasplatform.uk/bs-select/bs-select-app/-/merge_requests/944)
in `bs-select-app`
- [Jira issue](https://nhsd-jira.digital.nhs.uk/browse/BSS2-2403)